+++ src/nvlib/nv_globals.py
@@ -5,4 +5,4 @@ For further information see https://github.com/peter88213/
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 prefs = {}
-launchers = {}
+