+++ src/nvlib/plugin/plugin_collection.py
@@ -54,7 +54,7 @@ class PluginCollection(dict):
         except ValueError:
             # Set defaults for testing.
             self.majorVersion = 3
-            self.minorVersion = 0
+            self.minorVersion = 4
             self.patchlevel = 0
 
     def delete_file(self, moduleName):