+++ src/nvlib/view/properties_window/world_element_view.py
@@ -10,7 +10,6 @@ from tkinter import ttk
 
 from novxlib.novx_globals import _
 from novxlib.novx_globals import list_to_string
-from novxlib.novx_globals import Error
 from novxlib.novx_globals import string_to_list
 from nvlib.nv_globals import prefs
 from nvlib.view.properties_window.basic_view import BasicView
@@ -187,13 +186,7 @@ class WorldElementView(BasicView, ABC):
             return
 
         linkPath = list(self._element.links)[selection]
-        try:
-            self._mdl.linkProcessor.open_link(linkPath)
-        except Error as ex:
-            self._ui.show_error(
-                str(ex),
-                title=_('Cannot open link')
-                )
+        self._ctrl.open_link(linkPath)
 
     def _remove_link(self, event=None):
         """Remove a link from the list."""